call by value -> call by reference
A function may pass an argument by value, in which case invoking the function causes the arguments to be copied.  A function may pass an argument by reference, i.e., a pointer to the argument.  The latter saves copying, and is thus usually faster if the argument is a struct.  Throughout call by value has been replaced by call by reference.