set_sizes -> watch_list
Instead of having a vector of set_sizes for the various schools, we implement in terms of a list of sets of schools that have been found to be supercritical, so we can watch for them and then subdivide when the go critical.  This version of the code contains both implementations, so we are committing in preparation for stripping the old implementation out.