Bug removed
A problem with purify was that it could concentrate all roundoff error on a single student, whose allocation was then judged to not be an integer.
This was fixed by sharpening the tolerance in critpair, pushrelabel, and gcpscode from one in a million to one is a billion.  In addition, purifycode was rewritten to be drastically simpler.  Finally, a check was added to gcps.c, generating a sample product of purify and checking that it is efficient.