purify fixed and normal tweaked
purify was seriously flawed, in measuring the distance to a boundary element, and in setting the probabilities of the two directions.  Extensive debugging code led to a clear view of what was happening, which is correct after things were fixed.

normal() was sometimes returning inf or -inf if the first random integer was zero.  This has now been excluded.