bug fix, da executable added
There was a major bug, insofar as gcpsa and gcpsb were trying to use DA to compute a feasible allocation.  Encountering an example when this doesn't work, we eventually figured out that we need to use the allocations produced by fda.

We also added deferred acceptance as an executable da.